TeamCityTasks/StringExtensions.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;

namespace TeamCityTasks
{
    public static class StringExtensions
    {
        /// <summary>
        /// Escaped a value according TeamCity system messages rules.
        /// </summary>
        /// <see cref="https://confluence.jetbrains.com/display/TCD8/Build+Script+Interaction+with+TeamCity#BuildScriptInteractionwithTeamCity-servMsgsServiceMessages"/>
        public static string TeamCityEscape(this string value)
        {
            var sb = new StringBuilder(value.Length);
            foreach (var c in value)
            {
                switch (c)
                {
                    case '|':
                        sb.Append("||");
                        break;
                    case '\'':
                        sb.Append("|'");
                        break;
                    case '\n':
                        sb.Append("|n");
                        break;
                    case '\r':
                        sb.Append("|r");
                        break;
                    case '[':
                        sb.Append("|[");
                        break;
                    case ']':
                        sb.Append("|]");
                        break;
                    default:
                        sb.Append(c);
                        break;
                }
            }

            return sb.ToString();
        }
    }
}
